// data/github/current/ui/packages/billing-app/components/pickers/__generated__/UserPickerQuery.graphql.ts
export type UserPickerQuery$variables = {
  query?: string | null | undefined;
  slug: string;
};
export type UserPickerQuery$data = {
  readonly enterprise: {
    readonly members: {
      readonly nodes: ReadonlyArray<{
        readonly user?: {
          readonly " $fragmentSpreads": FragmentRefs<"UserPickerUserFragment">;
        } | null | undefined;
      } | null | undefined> | null | undefined;
    };
  } | null | undefined;
};
export type UserPickerQuery = {
  response: UserPickerQuery$data;
  variables: UserPickerQuery$variables;
};