// data/github/current/ui/packages/billing-app/components/pickers/__generated__/UserPickerEnterpriseOwnersQuery.graphql.ts
export type UserPickerEnterpriseOwnersQuery$variables = {
  slug: string;
};
export type UserPickerEnterpriseOwnersQuery$data = {
  readonly enterprise: {
    readonly ownerInfo: {
      readonly admins: {
        readonly nodes: ReadonlyArray<{
          readonly " $fragmentSpreads": FragmentRefs<"UserPickerUserFragment">;
        } | null | undefined> | null | undefined;
      };
    } | null | undefined;
  } | null | undefined;
};
export type UserPickerEnterpriseOwnersQuery = {
  response: UserPickerEnterpriseOwnersQuery$data;
  variables: UserPickerEnterpriseOwnersQuery$variables;
};